requirements.txt
textual==0.80.0
assemblyai==0.33.0
python-dotenv==1.0.1
pydantic==2.9.2
rich==13.9.2